+++ sites/m.pc6.com/components/mip-comment/index.less
@@ -1,155 +1,155 @@
 mip-comment {
-    fieldset {
-        border: 0
-    }
+  fieldset {
+    border: 0
+  }
 
-    input,
-    textarea {
-        outline: none;
-    }
+  input,
+  textarea {
+    outline: none;
+  }
 
-    #comment form {
-        padding: 20px 0;
-    }
+  #comment form {
+    padding: 20px 0;
+  }
 
-    #comment .d_title {
-        border-bottom: 1px solid #dfdfdf;
-        overflow: hidden;
-        line-height: 35px;
-        padding: 2px 10px 0 10px;
-        position: relative;
-    }
+  #comment .d_title {
+    border-bottom: 1px solid #dfdfdf;
+    overflow: hidden;
+    line-height: 35px;
+    padding: 2px 10px 0 10px;
+    position: relative;
+  }
 
-    #comment .d_title em {
-        font-size: 16px;
-        font-style: normal;
-        padding: 0 !important
-    }
+  #comment .d_title em {
+    font-size: 16px;
+    font-style: normal;
+    padding: 0 !important
+  }
 
-    #comment .fb {
-        display: block;
-        border-radius: 2px;
-        border: 1px solid #ccc;
-        color: #333;
-        -webkit-box-sizing: border-box;
-        line-height: 28px;
-        padding: 10px 5px 10px 46px;
-        position: relative;
-        margin: 8px 10px 0 10px;
-        background: url(../../static/all_ico.png) 5px -963px no-repeat;
-        box-shadow: 0px 2px 1px #fff, inset 0px 1px 1px rgba(138, 138, 138, 0.2);
-    }
+  #comment .fb {
+    display: block;
+    border-radius: 2px;
+    border: 1px solid #ccc;
+    color: #333;
+    -webkit-box-sizing: border-box;
+    line-height: 28px;
+    padding: 10px 5px 10px 46px;
+    position: relative;
+    margin: 8px 10px 0 10px;
+    background: url(../../static/all_ico.png) 5px -963px no-repeat;
+    box-shadow: 0px 2px 1px #fff, inset 0px 1px 1px rgba(138, 138, 138, 0.2);
+  }
 
-    #comment-list {
-        margin: 10px;
-        font-size: 14px;
-    }
+  #comment-list {
+    margin: 10px;
+    font-size: 14px;
+  }
 
-    #comment-list li {
-        padding: 10px 2px;
-        border-top: 1px solid #e6e6e6;
-        line-height: 24px;
-    }
+  #comment-list li {
+    padding: 10px 2px;
+    border-top: 1px solid #e6e6e6;
+    line-height: 24px;
+  }
 
-    #comment-list .user {
-        color: #65bb0a;
-        padding-right: 100px;
-        position: relative;
-        margin-bottom: 2px;
-        height: 24px;
-        overflow: hidden;
-    }
+  #comment-list .user {
+    color: #65bb0a;
+    padding-right: 100px;
+    position: relative;
+    margin-bottom: 2px;
+    height: 24px;
+    overflow: hidden;
+  }
 
-    .w-text {
-        margin: 0 15px 15px;
-        padding: 8px 20px 0 0;
-    }
+  .w-text {
+    margin: 0 15px 15px;
+    padding: 8px 20px 0 0;
+  }
 
-    .w-text textarea {
-        color: #666;
-        background: #fff;
-        border: 1px solid #c5c5c5;
-        width: 100%;
-        font-size: 24px;
-        -webkit-box-shadow: 0 1px 4px rgba(0, 0, 0, .1) inset;
-        height: 80px;
-        padding: 5px 10px;
-        line-height: 32px;
-        box-sizing: content-box;
-        -webkit-box-sizing: content-box;
-        font-size: 14px;
-    }
+  .w-text textarea {
+    color: #666;
+    background: #fff;
+    border: 1px solid #c5c5c5;
+    width: 100%;
+    font-size: 24px;
+    -webkit-box-shadow: 0 1px 4px rgba(0, 0, 0, .1) inset;
+    height: 80px;
+    padding: 5px 10px;
+    line-height: 32px;
+    box-sizing: content-box;
+    -webkit-box-sizing: content-box;
+    font-size: 14px;
+  }
 
-    .w-button {
-        margin: 0 20px;
-    }
+  .w-button {
+    margin: 0 20px;
+  }
 
-    #comment .button {
-        display: inline-block;
-        color: #555;
-        height: 32px;
-        border: 1px solid #ccc;
-        background: url(../../static/all_ico.png) 0 -1090px repeat-x;
-        -webkit-box-shadow: 0 1px 1px rgba(0, 0, 0, .06);
-        -moz-box-shadow: 0 1px 1px rgba(0, 0, 0, .06);
-        box-shadow: 0 1px 1px rgba(0, 0, 0, .06);
-        cursor: pointer;
-        text-align: center;
-        line-height: 32px;
-        height: 32px;
-        font-size: 14px;
-        border-radius: 0;
-        padding: 0;
-        margin: 0 0 0 2%;
-        vertical-align: middle;
-    }
+  #comment .button {
+    display: inline-block;
+    color: #555;
+    height: 32px;
+    border: 1px solid #ccc;
+    background: url(../../static/all_ico.png) 0 -1090px repeat-x;
+    -webkit-box-shadow: 0 1px 1px rgba(0, 0, 0, .06);
+    -moz-box-shadow: 0 1px 1px rgba(0, 0, 0, .06);
+    box-shadow: 0 1px 1px rgba(0, 0, 0, .06);
+    cursor: pointer;
+    text-align: center;
+    line-height: 32px;
+    height: 32px;
+    font-size: 14px;
+    border-radius: 0;
+    padding: 0;
+    margin: 0 0 0 2%;
+    vertical-align: middle;
+  }
 
-    #submit .button {
-        width: 47%;
-        margin-left: 2%;
-    }
+  #submit .button {
+    width: 47%;
+    margin-left: 2%;
+  }
 
-    #comment #cancel {}
+  #comment #cancel {}
 
 
 
-    #comment .comment-toggle-enter #view-comment {
-        display: none
-    }
+  #comment .comment-toggle-enter #view-comment {
+    display: none
+  }
 
-    #comment #submit {
-        display: none
-    }
+  #comment #submit {
+    display: none
+  }
 
-    #comment .comment-toggle-enter #submit {
-        display: block
-    }
+  #comment .comment-toggle-enter #submit {
+    display: block
+  }
 
-    #view-comment .button-status-complete {
-        border-top: 1px solid #e6e6e6;
-        overflow: hidden;
-        text-align: center;
-        position: relative;
-        display: none;
-        padding: 0;
-        padding-top: 15px;
-    }
+  #view-comment .button-status-complete {
+    border-top: 1px solid #e6e6e6;
+    overflow: hidden;
+    text-align: center;
+    position: relative;
+    display: none;
+    padding: 0;
+    padding-top: 15px;
+  }
 
-    #view-comment .button {
-        padding: 0 20px;
-    }
+  #view-comment .button {
+    padding: 0 20px;
+  }
 
-    #submit #verify {
-        color: #fff;
-        border: 1px solid #94c804;
-        background-position: 0 -1055px;
-        cursor: pointer;
-    }
-    #submit #verify.disable {
-        background-position: 0 -1090px ;
-        color: #999;
-        height: 34px;
-        line-height: 34px;
-        border:1px #ccc solid;
-    }
+  #submit #verify {
+    color: #fff;
+    border: 1px solid #94c804;
+    background-position: 0 -1055px;
+    cursor: pointer;
+  }
+  #submit #verify.disable {
+    background-position: 0 -1090px ;
+    color: #999;
+    height: 34px;
+    line-height: 34px;
+    border:1px #ccc solid;
+  }
 }
