--- sites/m.pc6.com/components/mip-common/README.md
@@ -1,54 +1,0 @@
-# mip-common
-
-站点逻辑共用
-
-标题|内容
-----|----
-类型|通用
-支持布局|responsive, fixed-height, fill, container, fixed
-所需脚本| [https://c.mipcdn.com/extensions/platform/v2/mpc6.com/mip-common/mip-common.js](https://c.mipcdn.com/extensions/platform/v2/mpc6.com/mip-common/mip-common.js)
-
-## 示例
-
-### 基本使用
-
-```js
-import MIPCommon from '../mip-common/mip-common'
-```
-
-## 方法
-
-### setCookie
-
-说明：设置cookies
-
-#### 参数
-
-##### name
-说明:cookie key名称
-
-##### value
-说明:cookie值
-
-### getCookie
-
-说明：读取cookies
-
-#### 参数
-
-##### name
-说明:cookie key名称
-
-### time
-
-说明：时间函数
-
-#### 参数
-
-##### cAlert
-说明:提示框 取代MIP不支持的alert
-
-### msg
-
-说明：提示语
-
