+++ sites/m.pc6.com/components/mip-down-custom/index.less
@@ -1,50 +1,3 @@
 mip-down-custom {
 
-    /*alert 提示*/
-    #alert_tips {
-        position: fixed;
-        background: #000;
-        opacity: .8;
-        color: #fff;
-        width: 200px;
-        line-height: 50px;
-        left: 50%;
-        top: 50%;
-        margin: -25px 0 0 -100px;
-        text-align: center;
-        border-radius: 5px;
-        z-index: 1000000
-    }
-
-    @-webkit-keyframes fadeIn {
-        0% {
-            opacity: 0
-        }
-
-        to {
-            opacity: 1
-        }
-    }
-
-    @keyframes fadeIn {
-        0% {
-            opacity: 0
-        }
-
-        to {
-            opacity: 1
-        }
-    }
-
-    .fadeIn {
-        -webkit-animation-name: fadeIn;
-        animation-name: fadeIn
-    }
-
-    .animated {
-        -webkit-animation-duration: 1s;
-        -webkit-animation-fill-mode: both;
-        animation-duration: 1s;
-        animation-fill-mode: both
-    }
 }
