--- sites/m.pc6.com/components/mip-statistics/README.md
@@ -1,17 +1,0 @@
-# mip-statistics
-
-页面自定义统计集合
-
-标题|内容
-----|----
-类型|
-支持布局|
-所需脚本| [https://c.mipcdn.com/extensions/platform/v2/mpc6.com/mip-statistics/mip-statistics.js](https://c.mipcdn.com/extensions/platform/v2/mpc6.com/mip-statistics/mip-statistics.js)
-
-## 示例
-
-### 基本使用
-
-```html
-<mip-statistics></mip-statistics>
-```
