--- sites/m.pc6.com/components/mip-statistics/example/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html mip>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
-    <title>MIP page</title>
-    <link rel="canonical" href="对应的原页面地址">
-    <link rel="stylesheet" href="https://c.mipcdn.com/static/v2/mip.css">
-    <style mip-custom>
-    /* 自定义样式 */
-    </style>
-  </head>
-  <body>
-    <div>
-      <mip-statistics></mip-statistics>
-    </div>
-    <script src="https://c.mipcdn.com/static/v2/mip.js"></script>
-    <script src="/mip-statistics/mip-statistics.js"></script>
-  </body>
-</html>
