--- sites/m.pc6.com/components/mip-statistics/index.less
@@ -1,6 +1,0 @@
-mip-statistics {
-  .wrapper {
-    margin: 0 auto;
-    text-align: center;
-  }
-}
